requirements.txt
Flask
Peewee
flask_restful
paramiko
gspread
oauth2client
google-api-python-client
python-nmap
pyyaml